- Refactored jdbc into separate package
