+++ nube-jdbc-engine/src/main/java/io/nubespark/jdbc/package-info.java
@@ -1,0 +1,5 @@
+@ModuleGen(name = JDBCService.SERVICE_NAME, groupPackage = "io.nubespark.jdbc")
+
+package io.nubespark.jdbc;
+
+import io.vertx.codegen.annotations.ModuleGen;
