--- nube-jdbc-engine/src/main/java/io/nubespark/package-info.java
@@ -1,5 +1,0 @@
-@ModuleGen(name = JDBCService.SERVICE_NAME, groupPackage = "io.nubespark")
-
-package io.nubespark;
-
-import io.vertx.codegen.annotations.ModuleGen;
